Added event manager....
Added event manager, modulerised the game so not everything is dealt
with in the main. Also added the concept of sprites although this is
currently not working to deal with the rendering of individual squares
as opposed to the entire game.
Moved all testing files into their own file. Completed OpenCV aspect of
game. Began producing a sudoko generator.
Made keyboard manager to deal with keyboard input and the CPU manager to
deal with quitting etc.